+++ packages/moneystream-extension/src/popup/components/QRCode.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import QRCode from 'qrcode.react'
+
+export function QRCodeImport (props:any) {
+    return (
+        <QRCode value={props.address} />
+    )
+}